Rename init.ts to ui.ts
